# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Service Ticket Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get high-level ticket summary counts and KPIs
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
        - name: state
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/active-tickets:
    get:
      operationId: getActiveTickets
      tags: [dashboard]
      summary: Get active tickets with age, pending counts, and breakdowns
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
        - name: state
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: pageSize
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Active tickets data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveTicketsResult"
  /dashboard/closed-tickets:
    get:
      operationId: getClosedTickets
      tags: [dashboard]
      summary: Get closed tickets with TAT breakdown by activity stage
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
        - name: state
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: pageSize
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Closed tickets data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosedTicketsResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, isActive, permissions]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, ash, rsh, service_partner]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        managerId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Upload:
      type: object
      required: [id, filename, fileType, recordCount, uploadedAt, status]
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
          enum: [active_tickets, closed_tickets, mrf_data]
        recordCount:
          type: integer
        uploadedAt:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        errorMessage:
          type: ["string", "null"]
...
    DashboardSummary:
      type: object
      required: [totalActive, totalClosed, totalMrf, avgTatMinutes, pendingByAge, byState, byCategory, recentUploads]
      properties:
        totalActive:
          type: integer
        totalClosed:
          type: integer
        totalMrf:
          type: integer
        avgTatMinutes:
          type: ["number", "null"]
        pendingByAge:
          type: object
          required: [within7Days, within15Days, within30Days, above30Days]
          properties:
            within7Days:
              type: integer
            within15Days:
              type: integer
            within30Days:
              type: integer
            above30Days:
              type: integer
        byState:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        recentUploads:
          type: array
          items:
            $ref: "#/components/schemas/Upload"
    CountByLabel:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer
    ActiveTicketsResult:
      type: object
      required: [tickets, total, byProduct, byServicePartner, byAsh, byRsh, byAgeBucket]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/ActiveTicket"
        total:
          type: integer
        byProduct:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        byServicePartner:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        byAsh:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        byRsh:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
        byAgeBucket:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
    ActiveTicket:
      type: object
      required: [ticketId, createdOn, product, category, servicePartner, ash, rsh, state, status, ageDays, supportType, ticketType]
      properties:
        ticketId:
          type: string
        createdOn:
          type: string
        product:
          type: string
        category:
          type: string
        servicePartner:
          type: string
        ash:
          type: string
        rsh:
          type: string
        state:
          type: string
        status:
          type: string
        wipSubStage:
          type: ["string", "null"]
        ageDays:
          type: integer
        supportType:
          type: string
        ticketType:
          type: string
        customerName:
          type: string
    ClosedTicketsResult:
      type: object
      required: [tickets, total, avgTat, byServicePartner, byAsh, byRsh, byState, byTicketType]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/ClosedTicket"
        total:
          type: integer
        avgTat:
          type: ["number", "null"]
        byServicePartner:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byAsh:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byRsh:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byState:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byTicketType:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
    AvgTatByLabel:
      type: object
      required: [label, count, avgTatMinutes]
      properties:
        label:
          type: string
        count:
          type: integer
        avgTatMinutes:
          type: ["number", "null"]
    ClosedTicket:
      type: object
      required: [ticketId, createdOn, product, category, servicePartner, ash, rsh, state, closedDate, tatMinutes, supportType, ticketType, closureType]
      properties:
        ticketId:
          type: string
        createdOn:
          type: string
        product:
          type: string
        category:
          type: string
        servicePartner:
          type: string
        ash:
          type: string
        rsh:
          type: string
        state:
          type: string
        closedDate:
          type: string
        tatMinutes:
          type: ["number", "null"]
        supportType:
          type: string
        ticketType:
          type: string
        closureType:
          type: ["string", "null"]
        customerName:
          type: string
...