# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Service Ticket Reporting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from './countByLabel';
import type { DashboardSummaryPendingByAge } from './dashboardSummaryPendingByAge';
import type { Upload } from './upload';

export interface DashboardSummary {
  totalActive: number;
  totalClosed: number;
  totalMrf: number;
  /** @nullable */
  avgTatMinutes: number | null;
  pendingByAge: DashboardSummaryPendingByAge;
  byState: CountByLabel[];
  byCategory: CountByLabel[];
  recentUploads: Upload[];
}